% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_cars.R
\docType{data}
\name{jhu_cars}
\alias{jhu_cars}
\title{Motor Trend Cars}
\format{
A \code{tibble} of the data
}
\usage{
jhu_cars
}
\description{
Motor Trend Cars
}
\note{
See \code{\link{mtcars}} for the full
}
\keyword{datasets}
