% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_file.R
\name{jhu_file}
\alias{jhu_file}
\title{Read JHU Data File}
\usage{
jhu_file(...)
}
\arguments{
\item{...}{Arguments to pass to
\code{\link{system.file}}}
}
\value{
A character vector
}
\description{
Read JHU Data File
}
\examples{
jhu_file('kaggleCarAuction.csv')
}
