% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bike.R
\name{read_bike}
\alias{read_bike}
\alias{read_bike_wide}
\title{Read Bike Lanes data}
\usage{
read_bike(...)

read_bike_wide(...)
}
\arguments{
\item{...}{Additional arguments to pass to
\code{\link{read_csv}}}
}
\value{
A \code{tibble}
}
\description{
Read Bike Lanes data
}
\examples{
bike = read_bike()
}
