% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_circulator.R
\name{read_circulator}
\alias{read_circulator}
\title{Read Charm City Circulator data}
\usage{
read_circulator(...)
}
\arguments{
\item{...}{Additional arguments to pass to
\code{\link{read_csv}}}
}
\value{
A \code{tibble}
}
\description{
Read Charm City Circulator data
}
\examples{
circ = read_circulator()
}
