% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kaggle.R
\name{read_kaggle}
\alias{read_kaggle}
\title{Read Kaggle Car Data}
\usage{
read_kaggle(col_types = cols(VehBCost = col_double()), ...)
}
\arguments{
\item{col_types}{column types to pass to
\code{\link{read_csv}}}

\item{...}{Additional arguments to pass to
\code{\link{read_csv}}}
}
\value{
A \code{tibble}
}
\description{
Read Kaggle Car Data
}
\examples{
x = read_kaggle()
}
