% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_monuments.R
\name{read_monuments}
\alias{read_monuments}
\title{Read Baltimore Monuments data}
\usage{
read_monuments(...)
}
\arguments{
\item{...}{Additional arguments to pass to
\code{\link{read_csv}}}
}
\value{
A \code{tibble}
}
\description{
Read Baltimore Monuments data
}
\examples{
mon = read_monuments()
}
