% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mortality.R
\name{read_mortality}
\alias{read_mortality}
\title{Read Child Mortality Data}
\usage{
read_mortality(...)
}
\arguments{
\item{...}{Additional arguments to pass to
\code{\link{read_csv}}}
}
\value{
A \code{tibble}
}
\description{
Read Child Mortality Data
}
\examples{
mort = read_mortality()
}
