% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_salaries.R
\name{read_salaries}
\alias{read_salaries}
\title{Read Baltimore Salary data}
\usage{
read_salaries(...)
}
\arguments{
\item{...}{Additional arguments to pass to
\code{\link{read_csv}}}
}
\value{
A \code{tibble}
}
\description{
Read Baltimore Salary data
}
\examples{
sal = read_salaries()
}
