% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tb.R
\name{read_tb}
\alias{read_tb}
\title{Read Tuberculosis data}
\usage{
read_tb(...)
}
\arguments{
\item{...}{Additional arguments to pass to
\code{\link{read_excel}}}
}
\value{
A \code{tibble}
}
\description{
Read Tuberculosis data
}
\examples{
tb = read_tb()
}
