% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_yts.R
\name{read_yts}
\alias{read_yts}
\title{Read Youth Tobacco Survey (YTS)}
\usage{
read_yts(...)
}
\arguments{
\item{...}{Additional arguments to pass to
\code{\link{read_csv}}}
}
\value{
A \code{tibble}
}
\description{
Read Youth Tobacco Survey (YTS)
}
\examples{
x = read_yts
}
